landesc.[c,h]: full rewrite
The code was stupid and ugly. Almost each module uses your own
implementation of list. This could bring more bugs. Nevertheless, the
commit brings a new implementation of list, cleans the code (a few
function was never used) and code is using a new list.

See more in src/list.h. This will be only way how to use list in
iptraf-ng.

Signed-off-by: Nikola Pajkovsky <npajkovs@redhat.com>