packet_get(): workaround select() bug
This is in linux select(2) man page:
-------------
Under Linux, select() may report a socket file descriptor as "ready for
reading",  while  nevertheless a subsequent read blocks. This could for
example happen when data has arrived but  upon  examination  has  wrong
checksum  and is discarded. There may be other circumstances in which a
file descriptor is spuriously reported as ready.  Thus it may be  safer
to use O_NONBLOCK on sockets that should not block.
-------------
Fix it by using MSG_DONTWAIT in recvfrom() (sets non-blocking mode).

Signed-off-by: Vitezslav Samel <vitezslav@samel.cz>
Signed-off-by: Nikola Pajkovsky <npajkovs@redhat.com>