packet.c: packet_get() fix
Formerly we used to signalize we don't have any packet by setting
its length to zero. This has been broken since commit
"packet.c: define struct pkt_hdr and functions on it", because
we incorectly zero the length only in the case recvfrom() failed.

Fix this by zeroing in every case and only set it to the right
value in the case of succesful recvfrom().

Signed-off-by: Vitezslav Samel <vitezslav@samel.cz>
Signed-off-by: Nikola Pajkovsky <npajkovs@redhat.com>