build: add support for ncurses[w][5,6]
now iptraf-ng is able to build against any ncurses library.
By default, iptraf-ng tries to find installed ncurses-config binary file
and set flags.

On Debian system, it will fail if libncuses5-dev is not installed,
because ncurses5-config is installed with essential package ncurses-bin.

If you want to override any ncurses-config, use of the commands:

$ make NEEDS_NCURSESW5=YesPlease
$ make NEEDS_NCURSES6=YesPlease
$ make NEEDS_NCURSESW6=YesPlease

or define it in Makefile or in config.mak

Acked-by: Vitezslav Samel <vitezslav@samel.cz>
Signed-off-by: Nikola Pajkovsky <npajkovs@redhat.com>