ipmon(): use struct rate to print TCP flow rate
This slightly changes user interface behavior: formerly the rate
was computed only for the entry under the bar; now the rate
is computed every second for every TCP entry in the table, so we have
the (accurate) rate printed right after the bar change - there is
no waittime to compute it. But this brings more CPU cycles burnt.

If this proves to be intollerable, we could rewrite it to compute
the TCP flow rate only for the entry under the bar.

Signed-off-by: Vitezslav Samel <vitezslav@samel.cz>
Signed-off-by: Nikola Pajkovsky <npajkovs@redhat.com>