ifstats.c: print_if_entries(): start at first visible entry
BUGFIX: When scrolling over many interfaces, the scrolling doesn't
stop and eventually segfaults, because the table->lastvisible is not
set properly (bug was introduced in "ifstats(): optimization").

In print_if_entries() we need to start at table->firstvisible entry
and print only visible entries (according to window height).

The sideeffect of this function is to properly set table->lastvisible.

Signed-off-by: Vitezslav Samel <vitezslav@samel.cz>
Signed-off-by: Nikola Pajkovsky <npajkovs@redhat.com>