use XDP to count packets
Add a '-b' option to count packets using an eBPF program instead of
an AF_PACKET socket in the detailed interface statistics mode.
When enabled, an eBPF will be attached to the interface and populate
a map with protocol statistics.
Then, iptraf-ng will read the map every DEFAULT_UPDATE_DELAY msecs
and update the panel.

The advantage is that the AF_PACKET socket has a big overhead because
it has to clone every packet and copy it to userspace.
With a fast link this overhead will become noticeable: iptraf-ng drops
packets and network traffic is severely impacted.
XDP instead can handle millions of packets per seconds without noticeable
effect on traffic.

This was tested on a 10 Gbit link filled with little less than 14 millions
of packets per second. The classic implementation capped at 111k pps with
both ksoftirqd and iptraf-ng stuck at 100% CPU usage. Network traffic was
fluctuating between -50% and -66%.
With the XDP program, all the traffic was correctly counted, iptraf-ng
CPU usage was almost 0%, and ksoftirqd oscillating in the 2-4% range.
Network traffic was not affected at all.

An asciinema recording of the session is here: https://asciinema.org/a/207160

The drawback of the XDP implementation is that only incoming traffic
is seen, since XDP doesn't handle outgoing traffic.

Signed-off-by: Matteo Croce <mcroce@microsoft.com>